requirements.txt
# Core
boto3
pyyaml
requests

# Analytics
pandas
numpy

# Testing
pytest
moto[s3,sns,dynamodb,sts]
